Rearranged packages and removed deprecations
