+++ pyproject.toml
@@ -65,6 +65,7 @@ sqlalchemy = ["sqlalchemy"]
 
 [tool.pytest.ini_options]
 asyncio_mode = "auto"
+asyncio_default_fixture_loop_scope = "session"
 addopts = "-xvrs --color=yes"
 junit_family = "xunit2"
 testpaths = [
