+++ pyuow/clock.py
@@ -1,0 +1,5 @@
+import datetime
+
+
+def offset_naive_utcnow() -> datetime.datetime:
+    return datetime.datetime.now(datetime.timezone.utc).replace(tzinfo=None)
