+++ pyuow/entity/base.py
@@ -2,7 +2,8 @@ import typing as t
 from dataclasses import dataclass
 from datetime import datetime
 
-from ...types import MISSING
+from ..clock import offset_naive_utcnow
+from ..types import MISSING
 
 ENTITY_ID = t.TypeVar("ENTITY_ID", bound=t.Any)
 ENTITY_TYPE = t.TypeVar("ENTITY_TYPE", bound="Entity[t.Any]")
@@ -19,7 +20,7 @@ class AuditedEntity(Entity[ENTITY_ID]):
     updated_date: datetime = t.cast(datetime, MISSING)
 
     def __post_init__(self) -> None:
-        now = datetime.utcnow()
+        now = offset_naive_utcnow()
 
         if self.created_date is MISSING:
             object.__setattr__(self, "created_date", now)
